+++ src/Vulkan/Extensions/VK_INTEL_shader_integer_functions2.hs
@@ -55,7 +55,7 @@ import Vulkan.Core10.Enums.StructureType (StructureType(STRUCTURE_TYPE_PHYSICAL_
 -- 'Vulkan.Core10.Enums.StructureType.StructureType'
 data PhysicalDeviceShaderIntegerFunctions2FeaturesINTEL = PhysicalDeviceShaderIntegerFunctions2FeaturesINTEL
   { -- | @shaderIntegerFunctions2@ indicates that the implementation supports the
-    -- @ShaderIntegerFunctions2INTEL@ SPIR-V capability.
+    -- @IntegerFunctions2INTEL@ SPIR-V capability.
     shaderIntegerFunctions2 :: Bool }
   deriving (Typeable, Eq)
 #if defined(GENERIC_INSTANCES)