+++ changelog.md
@@ -4,6 +4,10 @@
 
 ## [3.6.12] - 2020-11-03
 
+  - Bump API version to v1.2.159
+  - Include enum aliases defined in `<enums>` sections, until now only those
+    enum aliases in `features` or `extensions` were used.
+
 ## [3.6.11.1] - 2020-11-01
 
 - Raise bound on base